fix: review loop 3 — P0-P4 fixes across 11 crates
Non-atomic store writes fixed (chat, contact, outbox) with single
unchecked_transaction. RFC 8620 correctness: notFound SetError for
missing IDs, invalidProperties for unknown patch keys, createdIds in
JmapResponse, ResultReference name validation, Message/changes updated
field type. kith-tslocal: 30s WhoIs TTL cache, null-byte guards.
kith-tui: SSE auto-reconnect with exponential backoff, load error
feedback, deque invariant. kithd: TLS cert expiry sidecar, outbox
worker supervision loop, startup logging. kith-attach: 200 MiB blob
limit, base_dir existence guard. kithctl: CleanupGuard for backup stubs.
kith-core: unix_secs_to_rfc3339→u64, StateChange PartialEq,
make_channel(0) panic guard. kith-store: db_err maps constraint
violations to KithError::Validation.

+++ crates/kith-store/src/contact.rs
@@ -451,8 +451,9 @@ fn row_to_contact(row: &rusqlite::Row<'_>) -> rusqlite::Result<ChatContact> {
         id: peer_user_id,
         login: peer_login,
         display_name,
-        first_seen_at: crate::util::unix_secs_to_rfc3339(first_seen_at),
-        last_seen_at: crate::util::unix_secs_to_rfc3339(last_seen_at),
+        // Timestamps stored in SQLite are guaranteed non-negative Unix seconds.
+        first_seen_at: crate::util::unix_secs_to_rfc3339(first_seen_at as u64),
+        last_seen_at: crate::util::unix_secs_to_rfc3339(last_seen_at as u64),
         blocked: blocked != 0,
     })
 }
